requirements.txt
langchain==0.3.7
langgraph==0.2.39
langchain-openai==0.2.3
loguru==0.7.2
python-dotenv==1.0.1